Fix migrations

+++ src/migrations/2015_11_23_311098_add_language_id_column_to_sessions.php
@@ -21,7 +21,7 @@ class AddLanguageIdColumnToSessions extends Migration
         $this->builder->table(
             $this->table,
             function ($table) {
-                $table->integer('language_id')->unsigned()->nullable()->index();
+                $table->bigInteger('language_id')->unsigned()->nullable()->index();
             }
         );
     }
